fix: форма смены пароля вынесена в отдельный компонент
